# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: DraftMeld API
  version: 0.3.0
  description: Public contract between the DraftMeld frontend and backend.
  license:
    name: GNU Affero General Public License v3.0
    identifier: AGPL-3.0-only
servers:
  - url: /api/v1
security: []
paths:
...
  /ranking-sources:
    get:
      operationId: listRankingSources
      summary: List built-in ranking feeds and their refresh status
      responses:
        "200":
          description: Ranking source catalog.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RankingSource" }
        "500": { $ref: "#/components/responses/ServerError" }
  /ranking-sources/refresh:
    post:
      operationId: refreshRankingSources
      summary: Download and normalize every built-in ranking feed
      responses:
        "200":
          description: Refreshed ranking sources.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RankingSource" }
        "502":
          description: An upstream ranking feed could not be downloaded or parsed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /ranking-sources/import-pdf:
    post:
      operationId: importRankingPDF
      summary: Detect and import a supported user-supplied ranking PDF
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: The detected ranking source was normalized and imported.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RankingPDFImport" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "413":
          description: The uploaded PDF exceeds the 20 MiB limit.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /projection-sources/import-csv:
    post:
      operationId: importProjectionCSV
      summary: Import granular player projections from CSV
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, file]
              properties:
                name: { type: string }
                file: { type: string, format: binary }
      responses:
        "201":
          description: Imported projection source.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectionSource" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /ranking-identities/review:
    post:
      operationId: reviewRankingIdentity
      summary: Record a human review of an identity issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [issueKey, resolution]
              properties:
                issueKey: { type: string }
                resolution: { type: string, enum: [confirmed-separate, acknowledged] }
      responses:
        "204": { description: Identity review saved. }
  /rankings:
    get:
      operationId: getConsensusRankings
      summary: Build a league-weighted consensus from every imported feed
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: Normalized consensus rankings.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ConsensusRanking" }
        "500": { $ref: "#/components/responses/ServerError" }
  /ranking-watchlist:
    get:
      operationId: getRankingWatchlist
      summary: Find players favored by sources excluded from the league consensus
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: Up to five meaningful disabled-source ranking outliers.
          content:
            application/json:
              schema:
                type: array
                maxItems: 5
                items: { $ref: "#/components/schemas/WatchlistPlayer" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }
  /draft:
    get:
      operationId: getDraft
      summary: Get the current draft board and recommendations
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: Current draft state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftSnapshot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /draft/actions:
    post:
      operationId: recordDraftAction
      summary: Draft a player to my team or mark the player taken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftActionRequest"
      responses:
        "200":
          description: Updated draft state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftSnapshot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Player is no longer available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /draft/undo:
    post:
      operationId: undoDraftAction
      summary: Undo the most recent active draft action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [leagueId]
              properties:
                leagueId:
                  type: string
      responses:
        "200":
          description: Updated draft state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftSnapshot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: There is no active action to undo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/ServerError"
  /draft/preferences:
    put:
      operationId: setDraftPlayerPreference
      summary: Mark a player as a target, avoid, or neutral
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [leagueId, playerId, preference]
              properties:
                leagueId: { type: string }
                playerId: { type: string }
                preference: { type: string, enum: [target, avoid, ""] }
      responses:
        "200":
          description: Updated draft state.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DraftSnapshot" }
  /draft/mock:
    post:
      operationId: simulateDraftToNextTurn
      summary: Simulate opponent picks until the user's next turn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [leagueId]
              properties:
                leagueId: { type: string }
      responses:
        "200":
          description: Draft state after simulated opponent picks.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DraftSnapshot" }
  /draft/sync/sleeper:
    post:
      operationId: syncSleeperDraft
      summary: Import current picks from a public Sleeper draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [leagueId, sleeperDraftId, rosterId]
              properties:
                leagueId: { type: string }
                sleeperDraftId: { type: string }
                rosterId: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Draft state after Sleeper synchronization.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DraftSnapshot" }
components:
  parameters:
    LeagueId:
      name: leagueId
      in: query
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: The request was not valid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: The request could not be completed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
...
    RankingSource:
      type: object
      additionalProperties: false
      required: [id, name, description, methodology, license, projectUrl, dataUrl, defaultWeight, importMode, role, recordCount]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        methodology: { type: string }
        license: { type: string }
        projectUrl: { type: string, format: uri }
        dataUrl: { type: string, format: uri }
        defaultWeight: { type: number, exclusiveMinimum: 0 }
        importMode: { type: string, enum: [download, pdf-upload] }
        role: { type: string, enum: [ranking, market, usage] }
        recordCount: { type: integer, minimum: 0 }
        refreshedAt: { type: [string, "null"], format: date-time }
        publishedAt: { type: string }
...
    RankingPDFImport:
      type: object
      additionalProperties: false
      required: [source, pageCount]
      properties:
        source: { $ref: "#/components/schemas/RankingSource" }
        pageCount: { type: integer, minimum: 1 }
    ConsensusRanking:
      type: object
      additionalProperties: false
      required: [playerKey, name, position, team, rank, score, sourceCount, sourceRanks, coverage, rankRange, confidence, method]
      properties:
        playerKey: { type: string }
        name: { type: string }
        position: { type: string, enum: [QB, RB, WR, TE, K, DST] }
        team: { type: string }
        rank: { type: integer, minimum: 1 }
        score: { type: number }
        sourceCount: { type: integer, minimum: 1 }
        sourceRanks:
          type: object
          additionalProperties: { type: integer, minimum: 1 }
        coverage: { type: number, minimum: 0, maximum: 1 }
        rankRange: { type: integer, minimum: 0 }
        confidence: { type: string, enum: [high, medium, low] }
        method: { type: string, enum: [weighted-average, weighted-median, trimmed-mean] }
    ProjectionSource:
      type: object
      additionalProperties: false
      required: [id, name, recordCount, importedAt]
      properties:
        id: { type: string }
        name: { type: string }
        recordCount: { type: integer, minimum: 1 }
        importedAt: { type: string, format: date-time }
...
    WatchlistSignal:
      type: object
      additionalProperties: false
      required: [sourceId, sourceName, sourceRank, spotsHigher]
      properties:
        sourceId: { type: string }
        sourceName: { type: string }
        sourceRank: { type: integer, minimum: 1 }
        spotsHigher: { type: integer }
    WatchlistPlayer:
      type: object
      additionalProperties: false
      required: [playerKey, name, position, team, consensusRank, signals]
      properties:
        playerKey: { type: string }
        name: { type: string }
        position: { type: string, enum: [QB, RB, WR, TE, K, DST] }
        team: { type: string }
        consensusRank: { type: [integer, "null"], minimum: 1 }
        signals:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/WatchlistSignal" }
    Player:
      type: object
      required: [id, name, nflTeam, position, byeWeek, overallRank, positionRank, adp, tier, projectedPoints, valueOverReplacement, confidence, rankRange, preference, auctionValue]
      properties:
        id: { type: string }
        name: { type: string }
        nflTeam: { type: string }
        position: { type: string, enum: [QB, RB, WR, TE, K, DST] }
        byeWeek: { type: integer }
        overallRank: { type: integer, minimum: 1 }
        positionRank: { type: integer, minimum: 1 }
        adp: { type: number }
        tier: { type: integer, minimum: 1 }
        projectedPoints: { type: number }
        valueOverReplacement: { type: number }
        confidence: { type: string }
        rankRange: { type: integer, minimum: 0 }
        preference: { type: string, enum: [target, avoid, ""] }
        auctionValue: { type: number, minimum: 0 }
    Recommendation:
      type: object
      required: [player, score, reasons]
      properties:
        player: { $ref: "#/components/schemas/Player" }
        score: { type: number }
        reasons:
          type: array
          items: { type: string }
    Pick:
      type: object
      required: [eventId, number, action, player, createdAt, cost]
      properties:
        eventId: { type: integer }
        number: { type: integer, minimum: 1 }
        action: { type: string, enum: [draft, taken] }
        player: { $ref: "#/components/schemas/Player" }
        createdAt: { type: string, format: date-time }
        cost: { type: number, minimum: 0 }
    DraftSnapshot:
      type: object
      required: [leagueId, leagueName, pickNumber, available, myTeam, history, recommendations, canUndo, dataMode, projectionCount, draftType, nextUserPick, auctionBudget, budgetRemaining, auctionInflation, isUserTurn]
      properties:
        leagueId: { type: string }
        leagueName: { type: string }
        pickNumber: { type: integer, minimum: 1 }
        available:
          type: array
          items: { $ref: "#/components/schemas/Player" }
        myTeam:
          type: array
          items: { $ref: "#/components/schemas/Player" }
        history:
          type: array
          items: { $ref: "#/components/schemas/Pick" }
        recommendations:
          type: array
          items: { $ref: "#/components/schemas/Recommendation" }
        canUndo: { type: boolean }
        dataMode: { type: string }
        projectionCount: { type: integer, minimum: 0 }
        draftType: { type: string, enum: [snake, linear, auction] }
        nextUserPick: { type: integer, minimum: 0 }
        auctionBudget: { type: number, minimum: 0 }
        budgetRemaining: { type: number }
        auctionInflation: { type: number, minimum: 0 }
        isUserTurn: { type: boolean }
    DraftActionRequest:
      type: object
      required: [leagueId, playerId, action]
      properties:
        leagueId: { type: string }
        playerId: { type: string }
        action: { type: string, enum: [draft, taken] }
        cost: { type: number, minimum: 0 }